1.1.1 Update

+++ processor/comm.php
@@ -5,6 +5,12 @@ $MUST_END=false;
 $gameRequest[3] = @mb_convert_encoding($gameRequest[3], 'UTF-8', 'UTF-8');
 
 if ($gameRequest[0] == "init") { // Reset responses if init sent (Think about this)
+    // avoid a rare case where skyrim briefly reverts to level 1 Prisoner during load
+    if ($gameRequest[2] == "10000000") {
+        error_log("Ignoring init with a gamets of 10000000.");
+        $MUST_END=true;
+        return;
+    }
     $now=time();
     $db->delete("eventlog", "gamets>{$gameRequest[2]}  ");
     $db->delete("eventlog", "localts>$now ");
@@ -178,17 +184,20 @@ if ($gameRequest[0] == "init") { // Reset responses if init sent (Think about th
 
 
 } elseif ($gameRequest[0] == "_uquest") {
-    error_reporting(E_ALL);
 
     $questParsedData = explode("@",$gameRequest[3]);
-    print_r($questParsedData);
+    
     if (!empty($questParsedData[0])) {
         $data=array(
-                'briefing' => $questParsedData[2],
-                'data' => $questParsedData[2]
+            'ts' => $gameRequest[1],
+            'gamets' => $gameRequest[2],
+            'localts' => time(),
+            'briefing' => $questParsedData[2],
+            'data' => $questParsedData[2],
+            'id_quest'=>$questParsedData[0]
         );
         
-        $db->updateRow('quests',$data," id_quest='{$questParsedData[0]}' ");
+        $db->insert('questlog',$data);
 
     }
     $MUST_END=true;
@@ -232,7 +241,7 @@ if ($gameRequest[0] == "init") { // Reset responses if init sent (Think about th
                 'speaker' => $speech["speaker"],
                 'speech' => $speech["speech"],
                 'location' => $speech["location"],
-                'companions'=>(isset($speech["companions"])&&is_array($speech["companions"]))?implode(",",$speech["companions"]):"",
+                'companions'=>(isset($speech["companions"])&&is_array($speech["companions"]))?implode(",",$speech["companions"]):DataBeingsInCloseRange(),
                 'sess' => 'pending',
                 'audios' => isset($speech["audios"])?$speech["audios"]:null,
                 'topic' => isset($speech["debug"])?$speech["debug"]:null,
@@ -272,6 +281,7 @@ if ($gameRequest[0] == "init") { // Reset responses if init sent (Think about th
     $MUST_END=true;
 
 } elseif ($gameRequest[0] == "contentbook") {
+    // This should be deprecated once version 1.2.0 is released
     $db->insert(
         'books',
         array(
@@ -330,6 +340,7 @@ if ($gameRequest[0] == "init") { // Reset responses if init sent (Think about th
             $MUST_END=true;
         } else {
             logEvent($gameRequest);
+            
         }
     } else
         $MUST_END=true;
@@ -426,7 +437,7 @@ if ($gameRequest[0] == "init") { // Reset responses if init sent (Think about th
     
 } elseif ($gameRequest[0] == "setconf") {
     
-    //logEvent($gameRequest);
+    // logEvent($gameRequest);
 
     $vars=explode("@",$gameRequest[3]);
     $db->delete("conf_opts", "id='".$db->escape($vars[0])."'");
@@ -441,21 +452,33 @@ if ($gameRequest[0] == "init") { // Reset responses if init sent (Think about th
     
     $MUST_END=true;
     
-} elseif (strpos($gameRequest[0], "info")===0) {    // info_whatever commands
+} elseif (strpos($gameRequest[0], "info")===0) {    // info_whatever requests
 
     logEvent($gameRequest);
 
     $MUST_END=true;
 
     
-} elseif (strpos($gameRequest[0], "addnpc")===0) {    // info_whatever commands
+} elseif (strpos($gameRequest[0], "addnpc")===0) {    // addnpc 
     logEvent($gameRequest);
     
-    if (!profile_exists($gameRequest[3]))
-        AddFirstTimeMet($gameRequest[3], $momentum, $gameRequest[2],$gameRequest[1]);
+    $splitNameBase=explode("@",$gameRequest[3]);
+    if (sizeof($splitNameBase)>1) {
+        $localName=$splitNameBase[0];
+        $baseProfile=$splitNameBase[1];
+    } else {
+        $localName=$splitNameBase[0];
+        $baseProfile="";
+    }
 
-    createProfile($gameRequest[3],[],false);
-    
+    if ($localName==$baseProfile)
+        $baseProfile="";
+
+    if (!profile_exists($localName))
+        AddFirstTimeMet($localName, $momentum, $gameRequest[2],$gameRequest[1]);
+
+    createProfile($localName,[],false,$baseProfile);
+    audit_log("comm.php addnpc $localName");
     $MUST_END=true;
     
     
@@ -478,6 +501,8 @@ if ($gameRequest[0] == "init") { // Reset responses if init sent (Think about th
         $historyData="";
         $lastPlace="";
         $lastListener="";
+        $lastDateTime = "";
+
         foreach (json_decode(DataSpeechJournal($GLOBALS["HERIKA_NAME"],50),true) as $element) {
           if ($element["listener"]=="The Narrator") {
                 continue;
@@ -496,8 +521,17 @@ if ($gameRequest[0] == "init") { // Reset responses if init sent (Think about th
           }
           else
             $place="";
+
+            if ($lastDateTime != substr($element["sk_date"], 0, 15)) {
+                $date = substr($element["sk_date"], 0, 10);
+                $time = substr($element["sk_date"], 11);
+                $dateTime = "(on date {$date} at {$time})";
+                $lastDateTime = substr($element["sk_date"], 0, 15); 
+            } else {
+                $dateTime = "";
+            }
       
-          $historyData.=trim("{$element["speaker"]}:".trim($element["speech"])." $listener $place").PHP_EOL;
+          $historyData.=trim("{$element["speaker"]}:".trim($element["speech"])." $listener $place $dateTime").PHP_EOL;
           
         }
